@@ -16,7 +16,6 @@ import java.awt.datatransfer.UnsupportedFlavorException;
 import java.awt.dnd.DnDConstants;
 import java.awt.dnd.DropTarget;
 import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
 import java.awt.event.InputEvent;
 import java.awt.event.KeyEvent;
 import java.awt.event.MouseAdapter;
@@ -32,6 +31,7 @@ import java.io.OutputStreamWriter;
 import java.io.PrintStream;
 import java.net.ServerSocket;
 import java.net.Socket;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
@@ -82,13 +82,11 @@ import com.digero.common.abctomidi.FileAndData;
 import com.digero.common.icons.IconLoader;
 import com.digero.common.midi.MidiConstants;
 import com.digero.common.midi.LotroSequencerWrapper;
-import com.digero.common.midi.SequencerEvent;
 import com.digero.common.midi.SequencerEvent.SequencerProperty;
 import com.digero.common.midi.SequencerWrapper;
 import com.digero.common.midi.VolumeTransceiver;
 import com.digero.common.util.ExtensionFileFilter;
 import com.digero.common.util.FileFilterDropListener;
-import com.digero.common.util.Listener;
 import com.digero.common.util.LotroParseException;
 import com.digero.common.util.ParseException;
 import com.digero.common.util.Util;
@@ -101,8 +99,7 @@ import com.digero.common.view.SongPositionLabel;
 import com.digero.common.view.TempoBar;
 
 @SuppressWarnings("serial")
-public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConstants, TrackListPanelCallback
-{
+public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConstants, TrackListPanelCallback {
 	private static final ExtensionFileFilter ABC_FILE_FILTER = new ExtensionFileFilter("ABC Files", "abc", "txt");
 	public static final String APP_NAME = "ABC Player";
 	private static final String APP_NAME_LONG = APP_NAME + " for The Lord of the Rings Online";
@@ -112,67 +109,53 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 	static Version APP_VERSION = new Version(0, 0, 0);
 
 	private static AbcPlayer mainWindow = null;
-	
+
 	private JMenu recentItems;
 	private Queue<String> recentQueue;
 	private int recentMaxItems = 11;
-	
+
 	private static ServerSocket serverSocket;
 
-	public static void main(String[] args)
-	{
-		
-		try
-		{
+	public static void main(String[] args) {
+
+		try {
 			Properties props = new Properties();
 			props.load(AbcPlayer.class.getResourceAsStream("version.txt"));
 			String versionString = props.getProperty("version.AbcPlayer");
 			if (versionString != null)
 				APP_VERSION = Version.parseVersion(versionString);
+		} catch (IOException ex) {
 		}
-		catch (IOException ex)
-		{
-		}
-		
-		if(!openPort()) {
+
+		if (!openPort()) {
 			sendArgsToPort(args);
 			return;
 		}
 
 		System.setProperty("sun.sound.useNewAudioEngine", "true");
 
-		try
-		{
+		try {
 			UIManager.setLookAndFeel(UIManager.getSystemLookAndFeelClassName());
-		}
-		catch (Exception e)
-		{
+		} catch (Exception e) {
 		}
 
 		mainWindow = new AbcPlayer();
 		mainWindow.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 		mainWindow.setVisible(true);
 		mainWindow.openSongFromCommandLine(args);
-		try
-		{
+		try {
 			ready();
-		}
-		catch (UnsatisfiedLinkError err)
-		{
+		} catch (UnsatisfiedLinkError err) {
 			// Ignore (we weren't started via WinRun4j)
 		}
 	}
 
 	public static native boolean isVolumeSupported();
 
-	private static boolean isVolumeSupportedSafe()
-	{
-		try
-		{
+	private static boolean isVolumeSupportedSafe() {
+		try {
 			return isVolumeSupported();
-		}
-		catch (UnsatisfiedLinkError err)
-		{
+		} catch (UnsatisfiedLinkError err) {
 			return false;
 		}
 	}
@@ -181,8 +164,7 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 
 	public static native void setVolume(float volume);
 
-	public static void onVolumeChanged()
-	{
+	public static void onVolumeChanged() {
 		if (mainWindow != null && mainWindow.volumeBar != null)
 			mainWindow.volumeBar.repaint();
 	}
@@ -191,24 +173,22 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 	public static native void ready();
 
 	/** A new activation (a.k.a. a file was opened) */
-	public static void activate(String[] args)
-	{
+	public static void activate(String[] args) {
 		/*
-		if (args != null && args.length > 0 && args[0] != null) {
-			System.out.println(" Processing file path ("+args[0].length()+" chars):\n"+args[0]);
-		}
-		*/
-		mainWindow.openSongFromCommandLine(args);
-	}
-	
-	/** A new activation from WinRun4J 64bit (a.k.a. a file was opened) */
-	public static void activate(String arg0) {
-		final String[] args = {arg0.substring(1, arg0.length()-1)};
+		 * if (args != null && args.length > 0 && args[0] != null) {
+		 * System.out.println(" Processing file path ("+args[0].length()+" chars):\n"
+		 * +args[0]); }
+		 */
 		mainWindow.openSongFromCommandLine(args);
 	}
 
-	public static void execute(String cmdLine)
-	{
+	/** A new activation from WinRun4J 64bit (a.k.a. a file was opened) */
+	public static void activate(String arg0) {
+		final String[] args = { arg0.substring(1, arg0.length() - 1) };
+		mainWindow.openSongFromCommandLine(args);
+	}
+
+	public static void execute(String cmdLine) {
 		mainWindow.openSongFromCommandLine(new String[] { cmdLine });
 	}
 
@@ -246,7 +226,7 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 
 	private HighlightAbcNotesFrame abcViewFrame;
 
-	private final Map<Integer, LotroInstrument> instrumentOverrideMap = new HashMap<Integer, LotroInstrument>();
+	private final Map<Integer, LotroInstrument> instrumentOverrideMap = new HashMap<>();
 	private List<FileAndData> abcData;
 	private AbcInfo abcInfo = new AbcInfo();
 
@@ -254,67 +234,53 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 
 	private boolean isExporting = false;
 
-	public AbcPlayer()
-	{
+	public AbcPlayer() {
 		super(APP_NAME);
 
-		Runtime.getRuntime().addShutdownHook(new Thread()
-		{
-			@Override public void run()
-			{
+		Runtime.getRuntime().addShutdownHook(new Thread() {
+			@Override
+			public void run() {
 				if (sequencer != null)
 					sequencer.close();
 			}
 		});
 
-		try
-		{
-			List<Image> icons = new ArrayList<Image>();
+		try {
+			List<Image> icons = new ArrayList<>();
 			icons.add(ImageIO.read(IconLoader.class.getResourceAsStream("abcplayer_16.png")));
 			icons.add(ImageIO.read(IconLoader.class.getResourceAsStream("abcplayer_32.png")));
 			setIconImages(icons);
-		}
-		catch (Exception ex)
-		{
+		} catch (Exception ex) {
 			// Ignore
 		}
 
 		dropListener = new FileFilterDropListener(true, "abc", "txt");
-		dropListener.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				FileFilterDropListener l = (FileFilterDropListener) e.getSource();
-				boolean append = (l.getDropEvent().getDropAction() == DnDConstants.ACTION_COPY);
-				SwingUtilities.invokeLater(new OpenSongRunnable(append, l.getDroppedFiles().toArray(new File[0])));
-			}
+		dropListener.addActionListener(e -> {
+			FileFilterDropListener l = (FileFilterDropListener) e.getSource();
+			boolean append = (l.getDropEvent().getDropAction() == DnDConstants.ACTION_COPY);
+			SwingUtilities.invokeLater(new OpenSongRunnable(append, l.getDroppedFiles().toArray(new File[0])));
 		});
 		new DropTarget(this, dropListener);
 
-		if (isVolumeSupportedSafe())
-		{
+		if (isVolumeSupportedSafe()) {
 			volumeTransceiver = null;
-		}
-		else
-		{
+		} else {
 			volumeTransceiver = new VolumeTransceiver();
 			volumeTransceiver.setVolume(prefs.getInt("volumizer", VolumeTransceiver.MAX_VOLUME));
 		}
-		volumeBar = new NativeVolumeBar(new NativeVolumeBar.Callback()
-		{
-			@Override public void setVolume(int volume)
-			{
+		volumeBar = new NativeVolumeBar(new NativeVolumeBar.Callback() {
+			@Override
+			public void setVolume(int volume) {
 				if (volumeTransceiver == null)
 					AbcPlayer.setVolume((float) volume / NativeVolumeBar.MAX_VOLUME);
-				else
-				{
+				else {
 					volumeTransceiver.setVolume(volume);
 					prefs.putInt("volumizer", volume);
 				}
 			}
 
-			@Override public int getVolume()
-			{
+			@Override
+			public int getVolume() {
 				if (volumeTransceiver == null)
 					return (int) (AbcPlayer.getVolume() * NativeVolumeBar.MAX_VOLUME);
 				else
@@ -322,37 +288,31 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 			}
 		});
 
-		try
-		{
-			if (useLotroInstruments)
-			{
+		try {
+			if (useLotroInstruments) {
 				sequencer = new LotroSequencerWrapper();
 
-				if (LotroSequencerWrapper.getLoadLotroSynthError() != null)
-				{
+				if (LotroSequencerWrapper.getLoadLotroSynthError() != null) {
 					Version requredJavaVersion = new Version(1, 7, 0, 0);
 					Version recommendedJavaVersion = new Version(1, 7, 0, 25);
 
 					JPanel errorMessage = new JPanel(new BorderLayout(0, 12));
-					errorMessage.add(new JLabel(
-							"<html><b>There was an error loading the LOTRO instrument sounds</b><br>"
+					errorMessage
+							.add(new JLabel("<html><b>There was an error loading the LOTRO instrument sounds</b><br>"
 									+ "Playback will use standard MIDI instruments instead<br>"
 									+ "(drums do not sound good in this mode).</html>"), BorderLayout.NORTH);
 
 					final String JAVA_URL = "http://www.java.com";
-					if (requredJavaVersion.compareTo(Version.parseVersion(System.getProperty("java.version"))) > 0)
-					{
+					if (requredJavaVersion.compareTo(Version.parseVersion(System.getProperty("java.version"))) > 0) {
 						JLabel update = new JLabel("<html>It is recommended that you install Java "
 								+ recommendedJavaVersion.getMinor() + " update " + recommendedJavaVersion.getRevision()
 								+ " or later.<br>" + "Get the latest version from <a href='" + JAVA_URL + "'>"
 								+ JAVA_URL + "</a>.</html>");
 						update.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
-						update.addMouseListener(new MouseAdapter()
-						{
-							@Override public void mouseClicked(MouseEvent e)
-							{
-								if (e.getButton() == MouseEvent.BUTTON1)
-								{
+						update.addMouseListener(new MouseAdapter() {
+							@Override
+							public void mouseClicked(MouseEvent e) {
+								if (e.getButton() == MouseEvent.BUTTON1) {
 									Util.openURL(JAVA_URL);
 								}
 							}
@@ -360,36 +320,32 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 						errorMessage.add(update, BorderLayout.CENTER);
 					}
 
-					errorMessage.add(
-							new JLabel("<html>Error details:<br>" + LotroSequencerWrapper.getLoadLotroSynthError()
-									+ "</html>"), BorderLayout.SOUTH);
+					errorMessage.add(new JLabel(
+							"<html>Error details:<br>" + LotroSequencerWrapper.getLoadLotroSynthError() + "</html>"),
+							BorderLayout.SOUTH);
 
 					JOptionPane.showMessageDialog(this, errorMessage, APP_NAME + " failed to load LOTRO instruments",
 							JOptionPane.ERROR_MESSAGE);
 
 					useLotroInstruments = false;
 				}
-			}
-			else
-			{
+			} else {
 				sequencer = new SequencerWrapper();
 			}
 
 			if (volumeTransceiver != null)
 				sequencer.addTransceiver(volumeTransceiver);
-		}
-		catch (MidiUnavailableException e)
-		{
+		} catch (MidiUnavailableException e) {
 			JOptionPane.showMessageDialog(this, e.getMessage(), "MIDI error", JOptionPane.ERROR_MESSAGE);
 			System.exit(1);
 
-			// This will never be hit, but convinces the compiler that 
+			// This will never be hit, but convinces the compiler that
 			// the sequencer field will never be uninitialized
 			throw new RuntimeException();
 		}
 
 		content = new JPanel(new TableLayout(//
-				new double[] { 4, FILL, 4 },//
+				new double[] { 4, FILL, 4 }, //
 				new double[] { PREFERRED, 0, FILL, 8, PREFERRED }));
 		setContentPane(content);
 
@@ -406,7 +362,7 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 
 		JPanel controlPanel = new JPanel(new TableLayout(//
 				new double[] { 4, SongPositionBar.SIDE_PAD, 0.5, 4, PREFERRED, 4, PREFERRED, 4, 0.5,
-					SongPositionBar.SIDE_PAD, 4, PREFERRED, 4 },//
+						SongPositionBar.SIDE_PAD, 4, PREFERRED, 4 }, //
 				new double[] { 4, PREFERRED, 4, PREFERRED, 4 }));
 
 		songPositionBar = new SongPositionBar(sequencer);
@@ -424,24 +380,12 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		playButton = new JButton(playIcon);
 		playButton.setDisabledIcon(playIconDisabled);
 		playButton.setEnabled(false);
-		playButton.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				playPause();
-			}
-		});
+		playButton.addActionListener(e -> playPause());
 
 		stopButton = new JButton(stopIcon);
 		stopButton.setDisabledIcon(stopIconDisabled);
 		stopButton.setEnabled(false);
-		stopButton.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				stop();
-			}
-		});
+		stopButton.addActionListener(e -> stop());
 
 		tempoBar = new TempoBar(sequencer);
 
@@ -466,20 +410,14 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		controlPanel.add(volumePanel, "8, 3, c, c");
 		controlPanel.add(barNumberLabel, "9, 3, 11, 3, r, t");
 
-		sequencer.addChangeListener(new Listener<SequencerEvent>()
-		{
-			@Override public void onEvent(SequencerEvent evt)
-			{
-				SequencerProperty p = evt.getProperty();
-				if (!p.isInMask(SequencerProperty.THUMB_POSITION_MASK))
-				{
-					updateButtonStates();
-				}
+		sequencer.addChangeListener(evt -> {
+			SequencerProperty p = evt.getProperty();
+			if (!p.isInMask(SequencerProperty.THUMB_POSITION_MASK)) {
+				updateButtonStates();
+			}
 
-				if (p.isInMask(SequencerProperty.TEMPO.mask | SequencerProperty.SEQUENCE.mask))
-				{
-					updateTempoLabel();
-				}
+			if (p.isInMask(SequencerProperty.TEMPO.mask | SequencerProperty.SEQUENCE.mask)) {
+				updateTempoLabel();
 			}
 		});
 
@@ -495,22 +433,21 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		Util.initWinBounds(this, prefs.node("window"), 450, 282);
 	}
 
-	@Override public void setTrackInstrumentOverride(int trackIndex, LotroInstrument instrument)
-	{
+	@Override
+	public void setTrackInstrumentOverride(int trackIndex, LotroInstrument instrument) {
 		instrumentOverrideMap.put(trackIndex, instrument);
 		refreshSequence();
 	}
 
-	@Override public void showHighlightPanelForTrack(int trackIndex)
-	{
+	@Override
+	public void showHighlightPanelForTrack(int trackIndex) {
 		updateAbcView(/* showIfHidden = */true, /* retainScrollPosition = */false);
 
 		abcViewFrame.scrollToLineNumber(abcInfo.getPartStartLine(trackIndex));
 		abcViewFrame.setFollowedTrackNumber(trackIndex);
 	}
 
-	private void updateAbcView(boolean showIfHidden, boolean retainScrollPosition)
-	{
+	private void updateAbcView(boolean showIfHidden, boolean retainScrollPosition) {
 		boolean hasAbcData = (abcData != null && abcInfo != null);
 
 		if (showAbcViewMenuItem != null)
@@ -520,28 +457,24 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		if (!isVisible && !showIfHidden)
 			return;
 
-		if (abcViewFrame == null)
-		{
+		if (abcViewFrame == null) {
 			abcViewFrame = new HighlightAbcNotesFrame(sequencer);
 			abcViewFrame.setTitle(getTitle());
 			abcViewFrame.setIconImages(getIconImages());
 			abcViewFrame.addDropListener(dropListener);
 			abcViewFrame.setShowFullPartName(showFullPartNameMenuItem.isSelected());
-			abcViewFrame.addWindowListener(new WindowAdapter()
-			{
-				@Override public void windowOpened(WindowEvent e)
-				{
-					if (showAbcViewMenuItem != null)
-					{
+			abcViewFrame.addWindowListener(new WindowAdapter() {
+				@Override
+				public void windowOpened(WindowEvent e) {
+					if (showAbcViewMenuItem != null) {
 						showAbcViewMenuItem.setSelected(true);
 						prefs.putBoolean("showAbcViewMenuItem", true);
 					}
 				}
 
-				@Override public void windowClosed(WindowEvent e)
-				{
-					if (showAbcViewMenuItem != null)
-					{
+				@Override
+				public void windowClosed(WindowEvent e) {
+					if (showAbcViewMenuItem != null) {
 						showAbcViewMenuItem.setSelected(false);
 						prefs.putBoolean("showAbcViewMenuItem", false);
 					}
@@ -549,14 +482,11 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 			});
 		}
 
-		if (!hasAbcData)
-		{
+		if (!hasAbcData) {
 			abcViewFrame.setFollowedTrackNumber(-1);
 			abcViewFrame.clearLinesAndRegions();
-		}
-		else
-		{
-			List<String> lines = new ArrayList<String>();
+		} else {
+			List<String> lines = new ArrayList<>();
 			for (FileAndData entry : abcData)
 				lines.addAll(entry.lines);
 
@@ -567,27 +497,22 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 				abcViewFrame.scrollToLineNumber(startLine);
 		}
 
-		if (showIfHidden)
-		{
+		if (showIfHidden) {
 			if (!isVisible)
 				abcViewFrame.setVisible(true);
-			
+
 			abcViewFrame.toFront();
 		}
 	}
 
-	private void updateTitleLabel()
-	{
+	private void updateTitleLabel() {
 		String title = abcInfo.getTitle();
 		String artist = abcInfo.getComposer_MaybeNull();
 
-		if (artist != null)
-		{
+		if (artist != null) {
 			titleLabel.setText("<html>" + title + "&ensp;<span style='font-size:12pt; font-weight:normal'>" + artist
 					+ "</span></html>");
-		}
-		else
-		{
+		} else {
 			titleLabel.setText(title);
 		}
 
@@ -597,15 +522,13 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		titleLabel.setToolTipText(tooltip);
 	}
 
-	private void updateTempoLabel()
-	{
+	private void updateTempoLabel() {
 		float tempo = sequencer.getTempoFactor();
 		int t = Math.round(tempo * 100);
 		tempoLabel.setText("Tempo: " + t + "%");
 	}
 
-	private void initMenu()
-	{
+	private void initMenu() {
 		JMenuBar mainMenu = new JMenuBar();
 		setJMenuBar(mainMenu);
 
@@ -615,82 +538,58 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		JMenuItem open = fileMenu.add(new JMenuItem("Open ABC file(s)..."));
 		open.setMnemonic(KeyEvent.VK_O);
 		open.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_O, InputEvent.CTRL_DOWN_MASK));
-		open.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				openSongDialog();
-			}
-		});
+		open.addActionListener(e -> openSongDialog());
 
 		JMenuItem openAppend = fileMenu.add(new JMenuItem("Append ABC file(s)..."));
 		openAppend.setMnemonic(KeyEvent.VK_D);
-		openAppend.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_O, InputEvent.CTRL_DOWN_MASK
-				| InputEvent.SHIFT_DOWN_MASK));
-		openAppend.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				appendSongDialog();
-			}
-		});
+		openAppend.setAccelerator(
+				KeyStroke.getKeyStroke(KeyEvent.VK_O, InputEvent.CTRL_DOWN_MASK | InputEvent.SHIFT_DOWN_MASK));
+		openAppend.addActionListener(e -> appendSongDialog());
 
 		final JMenuItem pasteMenuItem = fileMenu.add(new JMenuItem("Open from clipboard"));
 		pasteMenuItem.setMnemonic(KeyEvent.VK_P);
 		pasteMenuItem.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_V, InputEvent.CTRL_DOWN_MASK));
-		pasteMenuItem.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				ArrayList<File> files = new ArrayList<File>();
-				if (getFileListFromClipboard(files))
-				{
-					openSong(files.toArray(new File[files.size()]));
-					return;
-				}
-
-				ArrayList<String> lines = new ArrayList<String>();
-				if (getAbcDataFromClipboard(lines, false))
-				{
-					List<FileAndData> filesData = new ArrayList<FileAndData>();
-					filesData.add(new FileAndData(new File("[Clipboard]"), lines));
-					openSong(filesData);
-					return;
-				}
-
-				Toolkit.getDefaultToolkit().beep();
+		pasteMenuItem.addActionListener(e -> {
+			ArrayList<File> files = new ArrayList<>();
+			if (getFileListFromClipboard(files)) {
+				openSong(files.toArray(new File[0]));
+				return;
 			}
+
+			ArrayList<String> lines = new ArrayList<>();
+			if (getAbcDataFromClipboard(lines, false)) {
+				List<FileAndData> filesData = new ArrayList<>();
+				filesData.add(new FileAndData(new File("[Clipboard]"), lines));
+				openSong(filesData);
+				return;
+			}
+
+			Toolkit.getDefaultToolkit().beep();
 		});
 
 		final JMenuItem pasteAppendMenuItem = fileMenu.add(new JMenuItem("Append from clipboard"));
 		pasteAppendMenuItem.setMnemonic(KeyEvent.VK_N);
-		pasteAppendMenuItem.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_V, InputEvent.CTRL_DOWN_MASK
-				| InputEvent.SHIFT_DOWN_MASK));
-		pasteAppendMenuItem.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				ArrayList<File> files = new ArrayList<File>();
-				if (getFileListFromClipboard(files))
-				{
-					appendSong(files.toArray(new File[files.size()]));
-					return;
-				}
-
-				ArrayList<String> lines = new ArrayList<String>();
-				if (getAbcDataFromClipboard(lines, true))
-				{
-					List<FileAndData> data = new ArrayList<FileAndData>();
-					data.add(new FileAndData(new File("[Clipboard]"), lines));
-					appendSong(data);
-					return;
-				}
-
-				Toolkit.getDefaultToolkit().beep();
+		pasteAppendMenuItem.setAccelerator(
+				KeyStroke.getKeyStroke(KeyEvent.VK_V, InputEvent.CTRL_DOWN_MASK | InputEvent.SHIFT_DOWN_MASK));
+		pasteAppendMenuItem.addActionListener(e -> {
+			ArrayList<File> files = new ArrayList<>();
+			if (getFileListFromClipboard(files)) {
+				appendSong(files.toArray(new File[0]));
+				return;
 			}
+
+			ArrayList<String> lines = new ArrayList<>();
+			if (getAbcDataFromClipboard(lines, true)) {
+				List<FileAndData> data = new ArrayList<>();
+				data.add(new FileAndData(new File("[Clipboard]"), lines));
+				appendSong(data);
+				return;
+			}
+
+			Toolkit.getDefaultToolkit().beep();
 		});
-		
-		recentQueue=new LinkedList<>();
+
+		recentQueue = new LinkedList<>();
 		recentItems = new JMenu();
 		recentItems.setText("Recent files...");
 		recentPrefsRead();
@@ -700,68 +599,49 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 
 		final JMenuItem saveMenuItem = fileMenu.add(new JMenuItem("Save a copy as ABC..."));
 		saveMenuItem.setMnemonic(KeyEvent.VK_S);
-		saveMenuItem.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_S, InputEvent.CTRL_DOWN_MASK
-				| InputEvent.SHIFT_DOWN_MASK));
-		saveMenuItem.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				if (!sequencer.isLoaded())
-				{
-					Toolkit.getDefaultToolkit().beep();
-					return;
-				}
-				saveSongDialog();
+		saveMenuItem.setAccelerator(
+				KeyStroke.getKeyStroke(KeyEvent.VK_S, InputEvent.CTRL_DOWN_MASK | InputEvent.SHIFT_DOWN_MASK));
+		saveMenuItem.addActionListener(e -> {
+			if (!sequencer.isLoaded()) {
+				Toolkit.getDefaultToolkit().beep();
+				return;
 			}
+			saveSongDialog();
 		});
 
 		final JMenuItem exportMp3MenuItem = fileMenu.add(new JMenuItem("Save as MP3 file (LAME)..."));
 		exportMp3MenuItem.setMnemonic(KeyEvent.VK_M);
 		exportMp3MenuItem.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_E, InputEvent.CTRL_DOWN_MASK));
-		exportMp3MenuItem.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				if (!sequencer.isLoaded() || isExporting)
-				{
-					Toolkit.getDefaultToolkit().beep();
-					return;
-				}
-				exportMp3();
+		exportMp3MenuItem.addActionListener(e -> {
+			if (!sequencer.isLoaded() || isExporting) {
+				Toolkit.getDefaultToolkit().beep();
+				return;
 			}
+			exportMp3();
 		});
-		
+
 		final JMenuItem exportMp3MenuItemNew = fileMenu.add(new JMenuItem("Save as MP3 file (FFmpeg)..."));
-		//exportMp3MenuItemNew.setMnemonic(KeyEvent.VK_M);
-		//exportMp3MenuItemNew.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_E, InputEvent.CTRL_DOWN_MASK));
-		exportMp3MenuItemNew.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				if (!sequencer.isLoaded() || isExporting)
-				{
-					Toolkit.getDefaultToolkit().beep();
-					return;
-				}
-				exportMp3New();
+		// exportMp3MenuItemNew.setMnemonic(KeyEvent.VK_M);
+		// exportMp3MenuItemNew.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_E,
+		// InputEvent.CTRL_DOWN_MASK));
+		exportMp3MenuItemNew.addActionListener(e -> {
+			if (!sequencer.isLoaded() || isExporting) {
+				Toolkit.getDefaultToolkit().beep();
+				return;
 			}
+			exportMp3New();
 		});
 
 		final JMenuItem exportWavMenuItem = fileMenu.add(new JMenuItem("Save as Wave file..."));
 		exportWavMenuItem.setMnemonic(KeyEvent.VK_E);
-		exportWavMenuItem.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_E, InputEvent.CTRL_DOWN_MASK
-				| InputEvent.SHIFT_DOWN_MASK));
-		exportWavMenuItem.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				if (!sequencer.isLoaded() || isExporting)
-				{
-					Toolkit.getDefaultToolkit().beep();
-					return;
-				}
-				exportWav();
+		exportWavMenuItem.setAccelerator(
+				KeyStroke.getKeyStroke(KeyEvent.VK_E, InputEvent.CTRL_DOWN_MASK | InputEvent.SHIFT_DOWN_MASK));
+		exportWavMenuItem.addActionListener(e -> {
+			if (!sequencer.isLoaded() || isExporting) {
+				Toolkit.getDefaultToolkit().beep();
+				return;
 			}
+			exportWav();
 		});
 
 		fileMenu.addSeparator();
@@ -769,18 +649,11 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		JMenuItem exit = fileMenu.add(new JMenuItem("Exit"));
 		exit.setMnemonic(KeyEvent.VK_X);
 		exit.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_F4, InputEvent.ALT_DOWN_MASK));
-		exit.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				System.exit(0);
-			}
-		});
+		exit.addActionListener(e -> System.exit(0));
 
-		fileMenu.addMenuListener(new MenuListener()
-		{
-			@Override public void menuSelected(MenuEvent e)
-			{
+		fileMenu.addMenuListener(new MenuListener() {
+			@Override
+			public void menuSelected(MenuEvent e) {
 				boolean pasteEnabled = getFileListFromClipboard(null);
 				pasteMenuItem.setEnabled(pasteEnabled || getAbcDataFromClipboard(null, false));
 				pasteAppendMenuItem.setEnabled(pasteEnabled || getAbcDataFromClipboard(null, true));
@@ -792,13 +665,13 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 				exportMp3MenuItemNew.setEnabled(saveEnabled && !isExporting);
 			}
 
-			@Override public void menuDeselected(MenuEvent e)
-			{
+			@Override
+			public void menuDeselected(MenuEvent e) {
 				menuCanceled(e);
 			}
 
-			@Override public void menuCanceled(MenuEvent e)
-			{
+			@Override
+			public void menuCanceled(MenuEvent e) {
 				pasteMenuItem.setEnabled(true);
 				pasteAppendMenuItem.setEnabled(true);
 
@@ -808,52 +681,31 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 				exportMp3MenuItemNew.setEnabled(true);
 			}
 		});
-		
+
 		JMenu editMenu = mainMenu.add(new JMenu(" Edit "));
 		editMenu.setMnemonic(KeyEvent.VK_E);
 		JMenuItem select = new JMenuItem("Select all");
 		JMenuItem deselect = new JMenuItem("Deselect all");
 		editMenu.add(select);
 		editMenu.add(deselect);
-		select.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				trackListPanel.selectAll();
-			}
-		});
-		deselect.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				trackListPanel.deselectAll();
-			}
-		});
+		select.addActionListener(e -> trackListPanel.selectAll());
+		deselect.addActionListener(e -> trackListPanel.deselectAll());
 
 		JMenu toolsMenu = mainMenu.add(new JMenu(" Tools "));
 		toolsMenu.setMnemonic(KeyEvent.VK_T);
 
 		toolsMenu.add(lotroErrorsMenuItem = new JCheckBoxMenuItem("Ignore LOTRO-specific errors"));
 		lotroErrorsMenuItem.setSelected(prefs.getBoolean("ignoreLotroErrors", false));
-		lotroErrorsMenuItem.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				prefs.putBoolean("ignoreLotroErrors", lotroErrorsMenuItem.isSelected());
-			}
-		});
+		lotroErrorsMenuItem
+				.addActionListener(e -> prefs.putBoolean("ignoreLotroErrors", lotroErrorsMenuItem.isSelected()));
 
 		toolsMenu.add(stereoMenuItem = new JCheckBoxMenuItem("Stereo pan in multi-part songs"));
 		stereoMenuItem.setToolTipText("<html>Separates the parts of a multi-part song by <br>"
 				+ "panning them towards the left or right speaker.</html>");
 		stereoMenuItem.setSelected(prefs.getBoolean("stereoMenuItem", true));
-		stereoMenuItem.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				prefs.putBoolean("stereoMenuItem", stereoMenuItem.isSelected());
-				refreshSequence();
-			}
+		stereoMenuItem.addActionListener(e -> {
+			prefs.putBoolean("stereoMenuItem", stereoMenuItem.isSelected());
+			refreshSequence();
 		});
 
 		toolsMenu.addSeparator();
@@ -863,185 +715,141 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		trackListPanel.setShowFullPartName(showFullPartNameMenuItem.isSelected());
 		if (abcViewFrame != null)
 			abcViewFrame.setShowFullPartName(showFullPartNameMenuItem.isSelected());
-		showFullPartNameMenuItem.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				prefs.putBoolean("showFullPartNameMenuItem", showFullPartNameMenuItem.isSelected());
-				trackListPanel.setShowFullPartName(showFullPartNameMenuItem.isSelected());
-				if (abcViewFrame != null)
-					abcViewFrame.setShowFullPartName(showFullPartNameMenuItem.isSelected());
-			}
+		showFullPartNameMenuItem.addActionListener(e -> {
+			prefs.putBoolean("showFullPartNameMenuItem", showFullPartNameMenuItem.isSelected());
+			trackListPanel.setShowFullPartName(showFullPartNameMenuItem.isSelected());
+			if (abcViewFrame != null)
+				abcViewFrame.setShowFullPartName(showFullPartNameMenuItem.isSelected());
 		});
 
 		final JCheckBoxMenuItem showLineNumbersMenuItem = new JCheckBoxMenuItem("Show line numbers");
 		toolsMenu.add(showLineNumbersMenuItem);
 		showLineNumbersMenuItem.setSelected(prefs.getBoolean("showLineNumbersMenuItem", true));
 		trackListPanel.setShowLineNumbers(showLineNumbersMenuItem.isSelected());
-		showLineNumbersMenuItem.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				prefs.putBoolean("showLineNumbersMenuItem", showLineNumbersMenuItem.isSelected());
-				trackListPanel.setShowLineNumbers(showLineNumbersMenuItem.isSelected());
-			}
+		showLineNumbersMenuItem.addActionListener(e -> {
+			prefs.putBoolean("showLineNumbersMenuItem", showLineNumbersMenuItem.isSelected());
+			trackListPanel.setShowLineNumbers(showLineNumbersMenuItem.isSelected());
 		});
 
 		final JCheckBoxMenuItem showSoloButtonsMenuItem = new JCheckBoxMenuItem("Show track solo buttons");
 		toolsMenu.add(showSoloButtonsMenuItem);
 		showSoloButtonsMenuItem.setSelected(prefs.getBoolean("showSoloButtonsMenuItem", true));
 		trackListPanel.setShowSoloButtons(showSoloButtonsMenuItem.isSelected());
-		showSoloButtonsMenuItem.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				prefs.putBoolean("showSoloButtonsMenuItem", showSoloButtonsMenuItem.isSelected());
-				trackListPanel.setShowSoloButtons(showSoloButtonsMenuItem.isSelected());
-			}
+		showSoloButtonsMenuItem.addActionListener(e -> {
+			prefs.putBoolean("showSoloButtonsMenuItem", showSoloButtonsMenuItem.isSelected());
+			trackListPanel.setShowSoloButtons(showSoloButtonsMenuItem.isSelected());
 		});
 
 		final JCheckBoxMenuItem showInstrumentComboBoxesMenuItem = new JCheckBoxMenuItem("Show instrument pickers");
 		toolsMenu.add(showInstrumentComboBoxesMenuItem);
 		showInstrumentComboBoxesMenuItem.setSelected(prefs.getBoolean("showInstrumentComboBoxesMenuItem", true));
 		trackListPanel.setShowInstrumentComboBoxes(showInstrumentComboBoxesMenuItem.isSelected());
-		showInstrumentComboBoxesMenuItem.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				prefs.putBoolean("showInstrumentComboBoxesMenuItem", showInstrumentComboBoxesMenuItem.isSelected());
-				trackListPanel.setShowInstrumentComboBoxes(showInstrumentComboBoxesMenuItem.isSelected());
-			}
+		showInstrumentComboBoxesMenuItem.addActionListener(e -> {
+			prefs.putBoolean("showInstrumentComboBoxesMenuItem", showInstrumentComboBoxesMenuItem.isSelected());
+			trackListPanel.setShowInstrumentComboBoxes(showInstrumentComboBoxesMenuItem.isSelected());
 		});
 
 		toolsMenu.addSeparator();
 
 		JMenuItem about = toolsMenu.add(new JMenuItem("About " + APP_NAME + "..."));
 		about.setMnemonic(KeyEvent.VK_A);
-		about.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				AboutDialog.show(AbcPlayer.this, APP_NAME_LONG, APP_VERSION, APP_URL, "abcplayer_64.png");
-			}
-		});
+		about.addActionListener(
+				e -> AboutDialog.show(AbcPlayer.this, APP_NAME_LONG, APP_VERSION, APP_URL, "abcplayer_64.png"));
 
 		JMenu abcViewMenu = mainMenu.add(new JMenu(" ABC View "));
 		abcViewMenu.setMnemonic(KeyEvent.VK_A);
 
 		abcViewMenu.add(showAbcViewMenuItem = new JCheckBoxMenuItem("Show ABC text"));
-		showAbcViewMenuItem.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				if (showAbcViewMenuItem.isSelected())
-				{
-					updateAbcView(/* showIfHidden = */true, /* retainScrollPosition = */false);
-				}
-				else if (abcViewFrame != null)
-				{
-					abcViewFrame.setVisible(false);
-				}
+		showAbcViewMenuItem.addActionListener(e -> {
+			if (showAbcViewMenuItem.isSelected()) {
+				updateAbcView(/* showIfHidden = */true, /* retainScrollPosition = */false);
+			} else if (abcViewFrame != null) {
+				abcViewFrame.setVisible(false);
 			}
 		});
 	}
-	
-	private void recentAdd (final String fileNameToAdd) {
+
+	private void recentAdd(final String fileNameToAdd) {
 		boolean removed = false;
 		if (recentQueue.contains(fileNameToAdd)) {
 			removed = recentQueue.remove(fileNameToAdd);
 		}
 		recentQueue.add(fileNameToAdd);
-		
-		if(recentQueue.size() > recentMaxItems) {
-            removed = true;
-            recentQueue.remove();
-		}
-        
-        if(removed){
-            recentItems.removeAll();
 
-            for (final String string : recentQueue) {
-                JMenuItem item = new JMenuItem(recentFilenameFromPath(string));
-                item.addActionListener(new java.awt.event.ActionListener() {
-                    public void actionPerformed(java.awt.event.ActionEvent evt) {
-                        recentActionPerformed(evt, string);
-                    }               
-                });
-                recentItems.add(item);
-            }
-        } else {
-        	JMenuItem newRecent = new JMenuItem(recentFilenameFromPath(fileNameToAdd));
-            newRecent.addActionListener(new java.awt.event.ActionListener() {
-                public void actionPerformed(java.awt.event.ActionEvent evt) {
-                    recentActionPerformed(evt, fileNameToAdd);
-                }               
-            });
-            recentItems.add(newRecent);
-        }
-        recentPrefsWrite();
-	}
-	
-	private void recentRemove (final String fileNameToRemove) {
-		if (recentQueue.remove(fileNameToRemove)) {
+		if (recentQueue.size() > recentMaxItems) {
+			removed = true;
+			recentQueue.remove();
+		}
+
+		if (removed) {
 			recentItems.removeAll();
-            for (final String string : recentQueue) {
-                JMenuItem item = new JMenuItem(recentFilenameFromPath(string));
-                item.addActionListener(new java.awt.event.ActionListener() {
-                    public void actionPerformed(java.awt.event.ActionEvent evt) {
-                        recentActionPerformed(evt, string);
-                    }
-                });
-                recentItems.add(item);
-            }
+
+			for (final String string : recentQueue) {
+				JMenuItem item = new JMenuItem(recentFilenameFromPath(string));
+				item.addActionListener(evt -> recentActionPerformed(evt, string));
+				recentItems.add(item);
+			}
+		} else {
+			JMenuItem newRecent = new JMenuItem(recentFilenameFromPath(fileNameToAdd));
+			newRecent.addActionListener(evt -> recentActionPerformed(evt, fileNameToAdd));
+			recentItems.add(newRecent);
 		}
 		recentPrefsWrite();
 	}
-	
-	private void recentPrefsWrite () {
+
+	private void recentRemove(final String fileNameToRemove) {
+		if (recentQueue.remove(fileNameToRemove)) {
+			recentItems.removeAll();
+			for (final String string : recentQueue) {
+				JMenuItem item = new JMenuItem(recentFilenameFromPath(string));
+				item.addActionListener(evt -> recentActionPerformed(evt, string));
+				recentItems.add(item);
+			}
+		}
+		recentPrefsWrite();
+	}
+
+	private void recentPrefsWrite() {
 		Object[] recentArray = recentQueue.toArray();
 		for (int i = 0; i < recentMaxItems; i++) {
 			if (i < recentQueue.size()) {
-				prefs.put("recent."+i, (String) recentArray[i]);
+				prefs.put("recent." + i, (String) recentArray[i]);
 			} else {
-				prefs.remove("recent."+i);
+				prefs.remove("recent." + i);
 			}
 		}
 	}
-	
-	private void recentPrefsRead () {
+
+	private void recentPrefsRead() {
 		for (int i = 0; i < recentMaxItems; i++) {
-			String entry = prefs.get("recent."+i, null);
-			if (entry != null) recentQueue.add(entry);
+			String entry = prefs.get("recent." + i, null);
+			if (entry != null)
+				recentQueue.add(entry);
 		}
 		recentItems.removeAll();
 
-        for (final String string : recentQueue) {
-            JMenuItem item = new JMenuItem(recentFilenameFromPath(string));
-            item.addActionListener(new java.awt.event.ActionListener() {
-                public void actionPerformed(java.awt.event.ActionEvent evt) {
-                    recentActionPerformed(evt, string);
-                }
-            });
-            recentItems.add(item);
-        }
+		for (final String string : recentQueue) {
+			JMenuItem item = new JMenuItem(recentFilenameFromPath(string));
+			item.addActionListener(evt -> recentActionPerformed(evt, string));
+			recentItems.add(item);
+		}
 	}
-	
-	private String recentFilenameFromPath (String path) {
+
+	private String recentFilenameFromPath(String path) {
 		Path p = Paths.get(path);
 		String fileName = p.getFileName().toString();
 		return fileName;
 	}
-	
-	private void recentActionPerformed(ActionEvent evt, String title) {
-		File[] files = {new File(title)};
-        if (!openSong(files)) {
-        	// The file could not be opened, removing it from the recent list.
-        	recentRemove(title);
-        }
-    }
 
-	private boolean getAbcDataFromClipboard(ArrayList<String> data, boolean checkContents)
-	{
+	private void recentActionPerformed(ActionEvent evt, String title) {
+		File[] files = { new File(title) };
+		if (!openSong(files)) {
+			// The file could not be opened, removing it from the recent list.
+			recentRemove(title);
+		}
+	}
+
+	private boolean getAbcDataFromClipboard(ArrayList<String> data, boolean checkContents) {
 		if (data != null)
 			data.clear();
 
@@ -1050,16 +858,9 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 			return false;
 
 		String text;
-		try
-		{
+		try {
 			text = (String) xfer.getTransferData(DataFlavor.stringFlavor);
-		}
-		catch (UnsupportedFlavorException e)
-		{
-			return false;
-		}
-		catch (IOException e)
-		{
+		} catch (UnsupportedFlavorException | IOException e) {
 			return false;
 		}
 
@@ -1069,25 +870,19 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		StringTokenizer tok = new StringTokenizer(text, "\r\n");
 		int i = 0;
 		boolean isValid = !checkContents;
-		while (tok.hasMoreTokens())
-		{
+		while (tok.hasMoreTokens()) {
 			String line = tok.nextToken();
 
-			if (!isValid)
-			{
-				if (line.startsWith("X:") || line.startsWith("x:"))
-				{
+			if (!isValid) {
+				if (line.startsWith("X:") || line.startsWith("x:")) {
 					isValid = true;
 					if (data == null)
 						break;
-				}
-				else
-				{
+				} else {
 					String lineTrim = line.trim();
-					// If we find a line that's not a comment before the 
+					// If we find a line that's not a comment before the
 					// X: line, then this isn't an ABC file
-					if (lineTrim.length() > 0 && !lineTrim.startsWith("%"))
-					{
+					if (lineTrim.length() > 0 && !lineTrim.startsWith("%")) {
 						isValid = false;
 						break;
 					}
@@ -1108,9 +903,8 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		return isValid;
 	}
 
-	@SuppressWarnings("unchecked")//
-	private boolean getFileListFromClipboard(ArrayList<File> data)
-	{
+	@SuppressWarnings("unchecked") //
+	private boolean getFileListFromClipboard(ArrayList<File> data) {
 		if (data != null)
 			data.clear();
 
@@ -1119,26 +913,17 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 			return false;
 
 		List<File> fileList;
-		try
-		{
+		try {
 			fileList = (List<File>) xfer.getTransferData(DataFlavor.javaFileListFlavor);
-		}
-		catch (UnsupportedFlavorException e)
-		{
-			return false;
-		}
-		catch (IOException e)
-		{
+		} catch (UnsupportedFlavorException | IOException e) {
 			return false;
 		}
 
 		if (fileList.size() == 0)
 			return false;
 
-		for (File file : fileList)
-		{
-			if (!ABC_FILE_FILTER.accept(file))
-			{
+		for (File file : fileList) {
+			if (!ABC_FILE_FILTER.accept(file)) {
 				if (data != null)
 					data.clear();
 				return false;
@@ -1151,68 +936,57 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		return true;
 	}
 
-	private void initOpenFileDialog()
-	{
-		if (openFileDialog == null)
-		{
-			openFileDialog = new JFileChooser(prefs.get("openFileDialog.currentDirectory", Util
-					.getLotroMusicPath(false).getAbsolutePath()));
+	private void initOpenFileDialog() {
+		if (openFileDialog == null) {
+			openFileDialog = new JFileChooser(
+					prefs.get("openFileDialog.currentDirectory", Util.getLotroMusicPath(false).getAbsolutePath()));
 
 			openFileDialog.setMultiSelectionEnabled(true);
 			openFileDialog.setFileFilter(ABC_FILE_FILTER);
 		}
 	}
 
-	private void openSongDialog()
-	{
+	private void openSongDialog() {
 		initOpenFileDialog();
 
 		int result = openFileDialog.showOpenDialog(this);
-		if (result == JFileChooser.APPROVE_OPTION)
-		{
+		if (result == JFileChooser.APPROVE_OPTION) {
 			prefs.put("openFileDialog.currentDirectory", openFileDialog.getCurrentDirectory().getAbsolutePath());
 
 			openSong(openFileDialog.getSelectedFiles());
 		}
 	}
 
-	private void appendSongDialog()
-	{
-		if (this.abcData == null || this.abcData.size() == 0)
-		{
+	private void appendSongDialog() {
+		if (this.abcData == null || this.abcData.size() == 0) {
 			openSongDialog();
 			return;
 		}
 
 		initOpenFileDialog();
 		int result = openFileDialog.showOpenDialog(this);
-		if (result == JFileChooser.APPROVE_OPTION)
-		{
+		if (result == JFileChooser.APPROVE_OPTION) {
 			prefs.put("openFileDialog.currentDirectory", openFileDialog.getCurrentDirectory().getAbsolutePath());
 
 			appendSong(openFileDialog.getSelectedFiles());
 		}
 	}
 
-	private void saveSongDialog()
-	{
-		if (this.abcData == null || this.abcData.size() == 0)
-		{
+	private void saveSongDialog() {
+		if (this.abcData == null || this.abcData.size() == 0) {
 			Toolkit.getDefaultToolkit().beep();
 			return;
 		}
 
-		if (saveFileDialog == null)
-		{
-			saveFileDialog = new JFileChooser(prefs.get("saveFileDialog.currentDirectory", Util
-					.getLotroMusicPath(false).getAbsolutePath()));
+		if (saveFileDialog == null) {
+			saveFileDialog = new JFileChooser(
+					prefs.get("saveFileDialog.currentDirectory", Util.getLotroMusicPath(false).getAbsolutePath()));
 
 			saveFileDialog.setFileFilter(ABC_FILE_FILTER);
 		}
 
 		int result = saveFileDialog.showSaveDialog(this);
-		if (result == JFileChooser.APPROVE_OPTION)
-		{
+		if (result == JFileChooser.APPROVE_OPTION) {
 			prefs.put("saveFileDialog.currentDirectory", saveFileDialog.getCurrentDirectory().getAbsolutePath());
 
 			String fileName = saveFileDialog.getSelectedFile().getName();
@@ -1221,8 +995,7 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 				fileName += ".abc";
 
 			File saveFileTmp = new File(saveFileDialog.getSelectedFile().getParent(), fileName);
-			if (saveFileTmp.exists())
-			{
+			if (saveFileTmp.exists()) {
 				int res = JOptionPane.showConfirmDialog(this, "File " + fileName + " already exists. Overwrite?",
 						"Confirm Overwrite", JOptionPane.YES_NO_OPTION, JOptionPane.WARNING_MESSAGE);
 				if (res != JOptionPane.YES_OPTION)
@@ -1234,15 +1007,12 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		}
 	}
 
-	private boolean openSongFromCommandLine(String[] args)
-	{
+	private boolean openSongFromCommandLine(String[] args) {
 		mainWindow.setExtendedState(mainWindow.getExtendedState() & ~JFrame.ICONIFIED);
 
-		if (args.length > 0)
-		{
+		if (args.length > 0) {
 			File[] argFiles = new File[args.length];
-			for (int i = 0; i < args.length; i++)
-			{
+			for (int i = 0; i < args.length; i++) {
 				argFiles[i] = new File(args[i]);
 			}
 			return openSong(argFiles);
@@ -1250,19 +1020,17 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		return false;
 	}
 
-	private class OpenSongRunnable implements Runnable
-	{
+	private class OpenSongRunnable implements Runnable {
 		private File[] abcFiles;
 		private boolean append;
 
-		public OpenSongRunnable(boolean append, File... abcFiles)
-		{
+		public OpenSongRunnable(boolean append, File... abcFiles) {
 			this.append = append;
 			this.abcFiles = abcFiles;
 		}
 
-		@Override public void run()
-		{
+		@Override
+		public void run() {
 			if (append)
 				appendSong(abcFiles);
 			else
@@ -1270,19 +1038,14 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		}
 	}
 
-	private boolean openSong(File[] abcFiles)
-	{
-		List<FileAndData> data = new ArrayList<FileAndData>();
+	private boolean openSong(File[] abcFiles) {
+		List<FileAndData> data = new ArrayList<>();
 
-		try
-		{
-			for (File abcFile : abcFiles)
-			{
+		try {
+			for (File abcFile : abcFiles) {
 				data.add(new FileAndData(abcFile, AbcToMidi.readLines(abcFile)));
 			}
-		}
-		catch (IOException e)
-		{
+		} catch (IOException e) {
 			JOptionPane.showMessageDialog(this, e.getMessage(), "Failed to open file", JOptionPane.ERROR_MESSAGE);
 			return false;
 		}
@@ -1290,19 +1053,14 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		return openSong(data);
 	}
 
-	private boolean appendSong(File[] abcFiles)
-	{
-		List<FileAndData> data = new ArrayList<FileAndData>();
+	private boolean appendSong(File[] abcFiles) {
+		List<FileAndData> data = new ArrayList<>();
 
-		try
-		{
-			for (File abcFile : abcFiles)
-			{
+		try {
+			for (File abcFile : abcFiles) {
 				data.add(new FileAndData(abcFile, AbcToMidi.readLines(abcFile)));
 			}
-		}
-		catch (IOException e)
-		{
+		} catch (IOException e) {
 			JOptionPane.showMessageDialog(this, e.getMessage(), "Failed to open file", JOptionPane.ERROR_MESSAGE);
 			return false;
 		}
@@ -1310,8 +1068,7 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		return appendSong(data);
 	}
 
-	private boolean onLotroParseError(LotroParseException lpe)
-	{
+	private boolean onLotroParseError(LotroParseException lpe) {
 		JCheckBox checkBox = new JCheckBox("Ignore LOTRO-specific errors");
 		Object[] message = new Object[] { lpe.getMessage(), checkBox };
 		JOptionPane.showMessageDialog(this, message, "Error reading ABC file", JOptionPane.WARNING_MESSAGE);
@@ -1320,16 +1077,12 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		return checkBox.isSelected();
 	}
 
-	private boolean openSong(List<FileAndData> data)
-	{
+	private boolean openSong(List<FileAndData> data) {
 		boolean isSameFile = false;
-		if (abcData != null && abcData.size() == data.size())
-		{
+		if (abcData != null && abcData.size() == data.size()) {
 			isSameFile = true;
-			for (int i = 0; i < abcData.size(); i++)
-			{
-				if (!abcData.get(i).file.equals(data.get(i).file))
-				{
+			for (int i = 0; i < abcData.size(); i++) {
+				if (!abcData.get(i).file.equals(data.get(i).file)) {
 					isSameFile = false;
 					break;
 				}
@@ -1342,11 +1095,9 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		Sequence song = null;
 		AbcInfo info = new AbcInfo();
 		boolean retry;
-		do
-		{
+		do {
 			retry = false;
-			try
-			{
+			try {
 				AbcToMidi.Params params = new AbcToMidi.Params(data);
 				params.useLotroInstruments = useLotroInstruments;
 				params.abcInfo = info;
@@ -1354,20 +1105,13 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 				params.stereo = stereoMenuItem.isSelected();
 				params.generateRegions = true;
 				song = AbcToMidi.convert(params);
-			}
-			catch (LotroParseException e)
-			{
-				if (onLotroParseError(e))
-				{
+			} catch (LotroParseException e) {
+				if (onLotroParseError(e)) {
 					retry = lotroErrorsMenuItem.isSelected();
-				}
-				else
-				{
+				} else {
 					return false;
 				}
-			}
-			catch (ParseException e)
-			{
+			} catch (ParseException e) {
 				JOptionPane.showMessageDialog(this, e.getMessage(), "Error reading ABC", JOptionPane.ERROR_MESSAGE);
 				return false;
 			}
@@ -1382,19 +1126,15 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 			abcViewFrame.clearLinesAndRegions();
 		stop();
 
-		try
-		{
+		try {
 			sequencer.setSequence(song);
-		}
-		catch (InvalidMidiDataException e)
-		{
+		} catch (InvalidMidiDataException e) {
 			JOptionPane.showMessageDialog(this, e.getMessage(), "MIDI error", JOptionPane.ERROR_MESSAGE);
 			return false;
 		}
 
 		sequencer.setTempoFactor(1.0f);
-		for (int i = 0; i < CHANNEL_COUNT; i++)
-		{
+		for (int i = 0; i < CHANNEL_COUNT; i++) {
 			sequencer.setTrackMute(i, false);
 			sequencer.setTrackSolo(i, false);
 		}
@@ -1410,9 +1150,8 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		updateAbcView(/* showIfHidden = */false, /* retainScrollPosition = */isSameFile);
 
 		sequencer.start();
-		
-		for (FileAndData fileAndData : data)
-		{
+
+		for (FileAndData fileAndData : data) {
 			String fileName = fileAndData.file.getAbsolutePath();
 			recentAdd(fileName);
 		}
@@ -1420,10 +1159,8 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		return true;
 	}
 
-	private boolean appendSong(List<FileAndData> appendData)
-	{
-		if (this.abcData == null || this.abcData.size() == 0)
-		{
+	private boolean appendSong(List<FileAndData> appendData) {
+		if (this.abcData == null || this.abcData.size() == 0) {
 			return openSong(appendData);
 		}
 
@@ -1431,17 +1168,15 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		long position = sequencer.getPosition();
 		sequencer.stop(); // pause
 
-		List<FileAndData> data = new ArrayList<FileAndData>(abcData);
+		List<FileAndData> data = new ArrayList<>(abcData);
 		data.addAll(appendData);
 
 		Sequence song = null;
 		AbcInfo info = new AbcInfo();
 		boolean retry;
-		do
-		{
+		do {
 			retry = false;
-			try
-			{
+			try {
 				AbcToMidi.Params params = new AbcToMidi.Params(data);
 				params.useLotroInstruments = useLotroInstruments;
 				params.instrumentOverrideMap = instrumentOverrideMap;
@@ -1450,33 +1185,23 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 				params.stereo = stereoMenuItem.isSelected();
 				params.generateRegions = true;
 				song = AbcToMidi.convert(params);
-			}
-			catch (LotroParseException e)
-			{
-				if (onLotroParseError(e))
-				{
+			} catch (LotroParseException e) {
+				if (onLotroParseError(e)) {
 					retry = lotroErrorsMenuItem.isSelected();
-				}
-				else
-				{
+				} else {
 					return false;
 				}
-			}
-			catch (ParseException e)
-			{
+			} catch (ParseException e) {
 				String thisFile = appendData.size() == 1 ? "this file" : "these files";
 				String msg = e.getMessage() + "\n\nWould you like to close the current song and retry opening "
 						+ thisFile + "?";
 				int result = JOptionPane.showConfirmDialog(this, msg, "Error appending ABC", JOptionPane.YES_NO_OPTION,
 						JOptionPane.ERROR_MESSAGE);
-				if (result == JOptionPane.YES_OPTION)
-				{
+				if (result == JOptionPane.YES_OPTION) {
 					boolean success = openSong(appendData);
 					sequencer.setRunning(success && running);
 					return success;
-				}
-				else
-				{
+				} else {
 					return false;
 				}
 			}
@@ -1487,22 +1212,18 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 
 		int oldTrackCount = sequencer.getSequence().getTracks().length;
 
-		try
-		{
+		try {
 			sequencer.reset(false);
 			sequencer.setSequence(song);
 			sequencer.setPosition(position);
 			sequencer.setRunning(running);
-		}
-		catch (InvalidMidiDataException e)
-		{
+		} catch (InvalidMidiDataException e) {
 			JOptionPane.showMessageDialog(this, e.getMessage(), "MIDI error", JOptionPane.ERROR_MESSAGE);
 			return false;
 		}
 
 		// Make sure the new tracks are unmuted
-		for (int i = oldTrackCount; i < CHANNEL_COUNT; i++)
-		{
+		for (int i = oldTrackCount; i < CHANNEL_COUNT; i++) {
 			sequencer.setTrackMute(i, false);
 			sequencer.setTrackSolo(i, false);
 		}
@@ -1520,27 +1241,24 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		return true;
 	}
 
-	private void updateWindowTitle()
-	{
-		String fileNames = "";
+	private void updateWindowTitle() {
+		StringBuilder fileNames = new StringBuilder();
 		int c = 0;
-		for (FileAndData fd : abcData)
-		{
+		for (FileAndData fd : abcData) {
 			File f = fd.file;
 			if (++c > 1)
-				fileNames += ", ";
+				fileNames.append(", ");
 
-			if (c > 2)
-			{
-				fileNames += "...";
+			if (c > 2) {
+				fileNames.append("...");
 				break;
 			}
 
-			fileNames += f.getName();
+			fileNames.append(f.getName());
 		}
 
 		String title = APP_NAME;
-		if (fileNames != "")
+		if (fileNames.toString() != "")
 			title += " - " + fileNames;
 
 		setTitle(title);
@@ -1548,52 +1266,37 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 			abcViewFrame.setTitle(title);
 	}
 
-	private boolean saveSong(File file)
-	{
-		PrintStream out = null;
-		try
-		{
-			out = new PrintStream(file);
+	private boolean saveSong(File file) {
+		try (PrintStream out = new PrintStream(file)) {
 			int i = 0;
-			for (FileAndData fileData : abcData)
-			{
+			for (FileAndData fileData : abcData) {
 				for (String line : fileData.lines)
 					out.println(line);
 
 				if (++i < abcData.size())
 					out.println();
 			}
-		}
-		catch (IOException e)
-		{
+		} catch (IOException e) {
 			JOptionPane.showMessageDialog(this, e.getMessage(), "Failed to save file", JOptionPane.ERROR_MESSAGE);
 			return false;
-		}
-		finally
-		{
-			if (out != null)
-				out.close();
 		}
 		return true;
 	}
 
-	private void playPause()
-	{
+	private void playPause() {
 		if (sequencer.isRunning())
 			sequencer.stop();
 		else
 			sequencer.start();
 	}
 
-	private void stop()
-	{
+	private void stop() {
 		sequencer.stop();
 		sequencer.setPosition(0);
 		updateButtonStates();
 	}
 
-	private void updateButtonStates()
-	{
+	private void updateButtonStates() {
 		boolean loaded = (sequencer.getSequence() != null);
 		playButton.setEnabled(loaded);
 		playButton.setIcon(sequencer.isRunning() ? pauseIcon : playIcon);
@@ -1601,10 +1304,8 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		stopButton.setEnabled(loaded && (sequencer.isRunning() || sequencer.getPosition() != 0));
 	}
 
-	private class WaitDialog extends JDialog
-	{
-		public WaitDialog(JFrame owner, File saveFile)
-		{
+	private class WaitDialog extends JDialog {
+		public WaitDialog(JFrame owner, File saveFile) {
 			super(owner, APP_NAME, false);
 			JPanel waitContent = new JPanel(new BorderLayout(5, 5));
 			setContentPane(waitContent);
@@ -1614,31 +1315,27 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 			waitProgress.setIndeterminate(true);
 			waitContent.add(waitProgress, BorderLayout.SOUTH);
 			pack();
-			setLocation(getOwner().getX() + (getOwner().getWidth() - getWidth()) / 2, getOwner().getY()
-					+ (getOwner().getHeight() - getHeight()) / 2);
+			setLocation(getOwner().getX() + (getOwner().getWidth() - getWidth()) / 2,
+					getOwner().getY() + (getOwner().getHeight() - getHeight()) / 2);
 			setResizable(false);
 			setEnabled(false);
 			setIconImages(AbcPlayer.this.getIconImages());
 		}
 	}
 
-	private void exportWav()
-	{
-		if (exportFileDialog == null)
-		{
-			exportFileDialog = new JFileChooser(prefs.get("exportFileDialog.currentDirectory", Util.getUserMusicPath()
-					.getAbsolutePath()));
+	private void exportWav() {
+		if (exportFileDialog == null) {
+			exportFileDialog = new JFileChooser(
+					prefs.get("exportFileDialog.currentDirectory", Util.getUserMusicPath().getAbsolutePath()));
 
 			File openedFile = null;
 			if (abcData.size() > 0)
 				openedFile = abcData.get(0).file;
 
-			if (openedFile != null)
-			{
+			if (openedFile != null) {
 				String openedName = openedFile.getName();
 				int dot = openedName.lastIndexOf('.');
-				if (dot >= 0)
-				{
+				if (dot >= 0) {
 					openedName = openedName.substring(0, dot);
 				}
 				openedName += ".wav";
@@ -1649,13 +1346,11 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		exportFileDialog.setFileFilter(new ExtensionFileFilter("WAV Files", "wav"));
 
 		int result = exportFileDialog.showSaveDialog(this);
-		if (result == JFileChooser.APPROVE_OPTION)
-		{
+		if (result == JFileChooser.APPROVE_OPTION) {
 			prefs.put("exportFileDialog.currentDirectory", exportFileDialog.getCurrentDirectory().getAbsolutePath());
 
 			File saveFile = exportFileDialog.getSelectedFile();
-			if (saveFile.getName().indexOf('.') < 0)
-			{
+			if (saveFile.getName().indexOf('.') < 0) {
 				saveFile = new File(saveFile.getParent() + "/" + saveFile.getName() + ".wav");
 				exportFileDialog.setSelectedFile(saveFile);
 			}
@@ -1666,62 +1361,46 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		}
 	}
 
-	private class ExportWavTask implements Runnable
-	{
+	private class ExportWavTask implements Runnable {
 		private Sequence sequence;
 		private File file;
 		private JDialog waitFrame;
 
-		public ExportWavTask(Sequence sequence, File file, JDialog waitFrame)
-		{
+		public ExportWavTask(Sequence sequence, File file, JDialog waitFrame) {
 			this.sequence = sequence;
 			this.file = file;
 			this.waitFrame = waitFrame;
 		}
 
-		@Override public void run()
-		{
+		@Override
+		public void run() {
 			isExporting = true;
 			Exception error = null;
-			try
-			{
-				FileOutputStream fos = new FileOutputStream(file);
-				try
-				{
+			try {
+				try (FileOutputStream fos = new FileOutputStream(file)) {
 					MidiToWav.render(sequence, fos);
 				}
-				finally
-				{
-					fos.close();
-				}
-			}
-			catch (Exception e)
-			{
+			} catch (Exception e) {
 				error = e;
-			}
-			finally
-			{
+			} finally {
 				isExporting = false;
 				SwingUtilities.invokeLater(new ExportWavFinishedTask(error, waitFrame));
 			}
 		}
 	}
 
-	private class ExportWavFinishedTask implements Runnable
-	{
+	private class ExportWavFinishedTask implements Runnable {
 		private Exception error;
 		private JDialog waitFrame;
 
-		public ExportWavFinishedTask(Exception error, JDialog waitFrame)
-		{
+		public ExportWavFinishedTask(Exception error, JDialog waitFrame) {
 			this.error = error;
 			this.waitFrame = waitFrame;
 		}
 
-		@Override public void run()
-		{
-			if (error != null)
-			{
+		@Override
+		public void run() {
+			if (error != null) {
 				JOptionPane.showMessageDialog(AbcPlayer.this, error.getMessage(), "Error saving WAV file",
 						JOptionPane.ERROR_MESSAGE);
 			}
@@ -1732,25 +1411,20 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 	// Cached result of isLame()
 	private static File notLameExe = null;
 
-	private boolean isLame(File lameExe)
-	{
+	private boolean isLame(File lameExe) {
 		if (!lameExe.exists() || lameExe.equals(notLameExe))
 			return false;
 
 		LameChecker checker = new LameChecker(lameExe);
 		checker.start();
-		try
-		{
+		try {
 			// Wait up to 3 seconds for the program to respond
 			checker.join(3000);
-		}
-		catch (InterruptedException e)
-		{
+		} catch (InterruptedException e) {
 		}
 		if (checker.isAlive())
 			checker.process.destroy();
-		if (!checker.isLame)
-		{
+		if (!checker.isLame) {
 			notLameExe = lameExe;
 			return false;
 		}
@@ -1758,61 +1432,49 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		return true;
 	}
 
-	private static class LameChecker extends Thread
-	{
+	private static class LameChecker extends Thread {
 		private boolean isLame = false;
 		private File lameExe;
 		private Process process;
 
-		public LameChecker(File lameExe)
-		{
+		public LameChecker(File lameExe) {
 			this.lameExe = lameExe;
 		}
 
-		@Override public void run()
-		{
-			try
-			{
-				process = Runtime.getRuntime().exec(new String[] {Util.quote(lameExe.getAbsolutePath()), " -?"});
+		@Override
+		public void run() {
+			try {
+				process = Runtime.getRuntime().exec(new String[] { Util.quote(lameExe.getAbsolutePath()), " -?" });
 				BufferedReader rdr = new BufferedReader(new InputStreamReader(process.getErrorStream()));
 				String line;
-				while ((line = rdr.readLine()) != null)
-				{
-					if (line.contains("LAME"))
-					{
+				while ((line = rdr.readLine()) != null) {
+					if (line.contains("LAME")) {
 						isLame = true;
 						break;
 					}
 				}
-			}
-			catch (IOException e)
-			{
+			} catch (IOException e) {
 			}
 		}
 	}
-	
+
 	private static File notFFExe = null;
 
-	private boolean isFF(File ffExe)
-	{
+	private boolean isFF(File ffExe) {
 		if (!ffExe.exists() || ffExe.equals(notFFExe))
 			return false;
 
 		FFChecker checker = new FFChecker(ffExe);
 		checker.start();
-		try
-		{
+		try {
 			// Wait up to 3 seconds for the program to respond
 			checker.join(3000);
-		}
-		catch (InterruptedException e)
-		{
+		} catch (InterruptedException e) {
 		}
 		if (checker.isAlive() && checker.process != null) {
 			checker.process.destroy();
 		}
-		if (!checker.isFF)
-		{
+		if (!checker.isFF) {
 			notFFExe = ffExe;
 			return false;
 		}
@@ -1820,61 +1482,48 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		return true;
 	}
 
-	private static class FFChecker extends Thread
-	{
+	private static class FFChecker extends Thread {
 		private boolean isFF = false;
 		private File ffExe;
 		private Process process;
 
-		public FFChecker(File ffExe)
-		{
+		public FFChecker(File ffExe) {
 			this.ffExe = ffExe;
 		}
 
-		@Override public void run()
-		{
-			try
-			{
-				process = Runtime.getRuntime().exec(new String[] {Util.quote(ffExe.getAbsolutePath()), "-help"});
+		@Override
+		public void run() {
+			try {
+				process = Runtime.getRuntime().exec(new String[] { Util.quote(ffExe.getAbsolutePath()), "-help" });
 				BufferedReader rdr = new BufferedReader(new InputStreamReader(process.getErrorStream()));
 				String line;
 				long start = System.currentTimeMillis();
-				while ((line = rdr.readLine()) != null || System.currentTimeMillis()-start < 50)
-				{
-					if (line != null && line.toLowerCase().contains("ffmpeg"))
-					{
-						//System.out.println("ERR:"+line);
+				while ((line = rdr.readLine()) != null || System.currentTimeMillis() - start < 50) {
+					if (line != null && line.toLowerCase().contains("ffmpeg")) {
+						// System.out.println("ERR:"+line);
 						isFF = true;
 						break;
 					} else if (line != null) {
-						//System.out.println("ERR:"+line);
+						// System.out.println("ERR:"+line);
 					}
 				}
-			}
-			catch (IOException e)
-			{
+			} catch (IOException e) {
 			}
 		}
 	}
 
-	private void exportMp3()
-	{
+	private void exportMp3() {
 		File openedFile = null;
 		if (abcData.size() > 0)
 			openedFile = abcData.get(0).file;
 
 		Preferences mp3Prefs = prefs.node("mp3");
 		File lameExe = new File(mp3Prefs.get("lameExe", "./lame.exe"));
-		if (!lameExe.exists())
-		{
-			outerLoop: for (File dir : new File(".").listFiles())
-			{
-				if (dir.isDirectory())
-				{
-					for (File file : dir.listFiles())
-					{
-						if (file.getName().toLowerCase().equals("lame.exe"))
-						{
+		if (!lameExe.exists()) {
+			outerLoop: for (File dir : new File(".").listFiles()) {
+				if (dir.isDirectory()) {
+					for (File file : dir.listFiles()) {
+						if (file.getName().equalsIgnoreCase("lame.exe")) {
 							lameExe = file;
 							break outerLoop;
 						}
@@ -1885,32 +1534,27 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 
 		JLabel hyperlink = new JLabel("<html><a href='" + LAME_URL + "'>" + LAME_URL + "</a></html>");
 		hyperlink.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
-		hyperlink.addMouseListener(new MouseAdapter()
-		{
-			@Override public void mouseClicked(MouseEvent e)
-			{
-				if (e.getClickCount() == 1 && e.getButton() == MouseEvent.BUTTON1)
-				{
+		hyperlink.addMouseListener(new MouseAdapter() {
+			@Override
+			public void mouseClicked(MouseEvent e) {
+				if (e.getClickCount() == 1 && e.getButton() == MouseEvent.BUTTON1) {
 					Util.openURL(LAME_URL);
 				}
 			}
 		});
 
 		boolean overrideAndUseExe = false;
-		for (int i = 0; (!overrideAndUseExe && !isLame(lameExe)) || !lameExe.exists(); i++)
-		{
-			if (i > 0)
-			{
+		for (int i = 0; (!overrideAndUseExe && !isLame(lameExe)) || !lameExe.exists(); i++) {
+			if (i > 0) {
 				JFileChooser fc = new JFileChooser();
-				fc.setFileFilter(new FileFilter()
-				{
-					@Override public boolean accept(File f)
-					{
-						return f.isDirectory() || f.getName().toLowerCase().equals("lame.exe");
+				fc.setFileFilter(new FileFilter() {
+					@Override
+					public boolean accept(File f) {
+						return f.isDirectory() || f.getName().equalsIgnoreCase("lame.exe");
 					}
 
-					@Override public String getDescription()
-					{
+					@Override
+					public String getDescription() {
 						return "lame.exe";
 					}
 				});
@@ -1924,19 +1568,15 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 				lameExe = fc.getSelectedFile();
 			}
 
-			if (!lameExe.exists())
-			{
+			if (!lameExe.exists()) {
 				Object message;
 				int icon;
-				if (i == 0)
-				{
+				if (i == 0) {
 					message = new Object[] {
-						"Exporting to MP3 requires LAME, a free MP3 encoder.\n" + "To download LAME, visit: ",
-						hyperlink, "\nAfter you download and unzip it, click OK to locate lame.exe", };
+							"Exporting to MP3 requires LAME, a free MP3 encoder.\n" + "To download LAME, visit: ",
+							hyperlink, "\nAfter you download and unzip it, click OK to locate lame.exe", };
 					icon = JOptionPane.INFORMATION_MESSAGE;
-				}
-				else
-				{
+				} else {
 					message = "File does not exist:\n" + lameExe.getAbsolutePath();
 					icon = JOptionPane.ERROR_MESSAGE;
 				}
@@ -1944,15 +1584,12 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 						JOptionPane.OK_CANCEL_OPTION, icon);
 				if (result != JOptionPane.OK_OPTION)
 					return;
-			}
-			else if (!isLame(lameExe))
-			{
+			} else if (!isLame(lameExe)) {
 				Object[] message = new Object[] {
-					"The MP3 converter you selected \"" + lameExe.getName() + "\" doesn't appear to be LAME.\n"
-							+ "You can download LAME from: ",
-					hyperlink,
-					"\nWould you like to use \"" + lameExe.getName() + "\" anyways?\n"
-							+ "If you choose No, you'll be prompted to locate lame.exe" };
+						"The MP3 converter you selected \"" + lameExe.getName() + "\" doesn't appear to be LAME.\n"
+								+ "You can download LAME from: ",
+						hyperlink, "\nWould you like to use \"" + lameExe.getName() + "\" anyways?\n"
+								+ "If you choose No, you'll be prompted to locate lame.exe" };
 				int result = JOptionPane.showConfirmDialog(this, message, "Export to MP3 requires LAME",
 						JOptionPane.YES_NO_CANCEL_OPTION, JOptionPane.WARNING_MESSAGE);
 				if (result == JOptionPane.YES_OPTION)
@@ -1969,66 +1606,49 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 		ExportMp3Dialog mp3Dialog = new ExportMp3Dialog(this, lameExe, mp3Prefs, openedFile, abcInfo.getTitle(),
 				abcInfo.getComposer());
 		mp3Dialog.setIconImages(AbcPlayer.this.getIconImages());
-		mp3Dialog.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				ExportMp3Dialog dialog = (ExportMp3Dialog) e.getSource();
-				JDialog waitFrame = new WaitDialog(AbcPlayer.this, dialog.getSaveFile());
-				waitFrame.setVisible(true);
-				new Thread(new ExportMp3Task(sequencer.getSequence(), dialog, waitFrame)).start();
-			}
+		mp3Dialog.addActionListener(e -> {
+			ExportMp3Dialog dialog = (ExportMp3Dialog) e.getSource();
+			JDialog waitFrame = new WaitDialog(AbcPlayer.this, dialog.getSaveFile());
+			waitFrame.setVisible(true);
+			new Thread(new ExportMp3Task(sequencer.getSequence(), dialog, waitFrame)).start();
 		});
 		mp3Dialog.setVisible(true);
 	}
 
-	private class ExportMp3Task implements Runnable
-	{
+	private class ExportMp3Task implements Runnable {
 		private Sequence sequence;
 		private ExportMp3Dialog mp3Dialog;
 		private JDialog waitFrame;
 
-		public ExportMp3Task(Sequence sequence, ExportMp3Dialog mp3Dialog, JDialog waitFrame)
-		{
+		public ExportMp3Task(Sequence sequence, ExportMp3Dialog mp3Dialog, JDialog waitFrame) {
 			this.sequence = sequence;
 			this.mp3Dialog = mp3Dialog;
 			this.waitFrame = waitFrame;
 		}
 
-		@Override public void run()
-		{
+		@Override
+		public void run() {
 			isExporting = true;
 			Exception error = null;
 			String lameExeSav = null;
 			Preferences mp3Prefs = mp3Dialog.getPreferencesNode();
-			try
-			{
+			try {
 				lameExeSav = mp3Prefs.get("lameExe", null);
 				mp3Prefs.put("lameExe", "");
 				File wavFile = File.createTempFile("AbcPlayer-", ".wav");
-				FileOutputStream fos = new FileOutputStream(wavFile);
-				try
-				{
+				try (FileOutputStream fos = new FileOutputStream(wavFile)) {
 					MidiToWav.render(sequence, fos);
 					fos.close();
 					Process p = Runtime.getRuntime().exec(mp3Dialog.getCommandLine(wavFile));
 					if (p.waitFor() != 0)
 						throw new Exception("LAME failed");
-				}
-				finally
-				{
-					fos.close();
+				} finally {
 					wavFile.delete();
 				}
-			}
-			catch (Exception e)
-			{
+			} catch (Exception e) {
 				error = e;
-			}
-			finally
-			{
-				if (lameExeSav != null)
-				{
+			} finally {
+				if (lameExeSav != null) {
 					mp3Prefs.put("lameExe", lameExeSav);
 				}
 				isExporting = false;
@@ -2036,218 +1656,195 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 			}
 		}
 	}
-	
-	private void exportMp3New()
-	{
+
+	private void exportMp3New() {
 		File openedFile = null;
 		if (abcData.size() > 0)
 			openedFile = abcData.get(0).file;
 
 		Preferences mp3Prefs = prefs.node("mp3");
 		File ffExe = new File(mp3Prefs.get("ffExe", "./ffmpeg.exe"));
-		if (!ffExe.exists())
-		{
-			outerLoop: for (File dir : new File(".").listFiles())
-			{
-				if (dir.isDirectory())
-				{
-					for (File file : dir.listFiles())
-					{
-						if (file.getName().toLowerCase().equals("ffmpeg.exe"))
-						{
-							ffExe = file;
-							break outerLoop;
+		if (System.getProperty("os.name").contains("Windows")) {
+
+			if (!ffExe.exists()) {
+				outerLoop: for (File dir : new File(".").listFiles()) {
+					if (dir.isDirectory()) {
+						for (File file : dir.listFiles()) {
+							if (file.getName().equalsIgnoreCase("ffmpeg.exe")) {
+								ffExe = file;
+								break outerLoop;
+							}
 						}
 					}
 				}
 			}
-		}
 
-		JLabel hyperlink = new JLabel("<html><a href='" + FF_URL + "'>" + FF_URL + "</a></html>");
-		hyperlink.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
-		hyperlink.addMouseListener(new MouseAdapter()
-		{
-			@Override public void mouseClicked(MouseEvent e)
-			{
-				if (e.getClickCount() == 1 && e.getButton() == MouseEvent.BUTTON1)
-				{
-					Util.openURL(FF_URL);
-				}
-			}
-		});
-
-		boolean overrideAndUseExe = false;
-		for (int i = 0; !ffExe.exists() || (!overrideAndUseExe && !isFF(ffExe)); i++)
-		{
-			if (i > 0)
-			{
-				JFileChooser fc = new JFileChooser();
-				fc.setFileFilter(new FileFilter()
-				{
-					@Override public boolean accept(File f)
-					{
-						return f.isDirectory() || f.getName().toLowerCase().equals("ffmpeg.exe");
+			JLabel hyperlink = new JLabel("<html><a href='" + FF_URL + "'>" + FF_URL + "</a></html>");
+			hyperlink.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
+			hyperlink.addMouseListener(new MouseAdapter() {
+				@Override
+				public void mouseClicked(MouseEvent e) {
+					if (e.getClickCount() == 1 && e.getButton() == MouseEvent.BUTTON1) {
+						Util.openURL(FF_URL);
 					}
+				}
+			});
 
-					@Override public String getDescription()
-					{
-						return "ffmpeg.exe";
+			boolean overrideAndUseExe = false;
+			for (int i = 0; !ffExe.exists() || (!overrideAndUseExe && !isFF(ffExe)); i++) {
+				if (i > 0) {
+					JFileChooser fc = new JFileChooser();
+					fc.setFileFilter(new FileFilter() {
+						@Override
+						public boolean accept(File f) {
+							return f.isDirectory() || f.getName().equalsIgnoreCase("ffmpeg.exe");
+						}
+
+						@Override
+						public String getDescription() {
+							return "ffmpeg.exe";
+						}
+					});
+					fc.setSelectedFile(ffExe);
+					int result = fc.showOpenDialog(this);
+
+					if (result == JFileChooser.ERROR_OPTION)
+						continue; // Try again
+					else if (result != JFileChooser.APPROVE_OPTION)
+						return;
+					ffExe = fc.getSelectedFile();
+				}
+
+				if (!ffExe.exists()) {
+					Object message;
+					int icon;
+					if (i == 0) {
+						message = new Object[] {
+								"Exporting to MP3 requires FFmpeg, a free MP3 encoder.\n"
+										+ "To download FFmpeg, visit: ",
+								hyperlink, "\nAfter you download and unzip it, click OK to locate ffmpeg.exe", };
+						icon = JOptionPane.INFORMATION_MESSAGE;
+					} else {
+						message = "File does not exist:\n" + ffExe.getAbsolutePath();
+						icon = JOptionPane.ERROR_MESSAGE;
 					}
-				});
-				fc.setSelectedFile(ffExe);
-				int result = fc.showOpenDialog(this);
-
-				if (result == JFileChooser.ERROR_OPTION)
-					continue; // Try again
-				else if (result != JFileChooser.APPROVE_OPTION)
-					return;
-				ffExe = fc.getSelectedFile();
-			}
-
-			if (!ffExe.exists())
-			{
-				Object message;
-				int icon;
-				if (i == 0)
-				{
-					message = new Object[] {
-						"Exporting to MP3 requires FFmpeg, a free MP3 encoder.\n" + "To download FFmpeg, visit: ",
-						hyperlink, "\nAfter you download and unzip it, click OK to locate ffmpeg.exe", };
-					icon = JOptionPane.INFORMATION_MESSAGE;
+					int result = JOptionPane.showConfirmDialog(this, message, "Export to MP3 requires FFmpeg",
+							JOptionPane.OK_CANCEL_OPTION, icon);
+					if (result != JOptionPane.OK_OPTION)
+						return;
+				} else if (!isFF(ffExe)) {
+					Object[] message = new Object[] {
+							"The MP3 converter you selected \"" + ffExe.getName() + "\" doesn't appear to be FFmpeg.\n"
+									+ "You can download FFmpeg from: ",
+							hyperlink, "\nWould you like to use \"" + ffExe.getName() + "\" anyways?\n"
+									+ "If you choose No, you'll be prompted to locate ffmpeg.exe" };
+					int result = JOptionPane.showConfirmDialog(this, message, "Export to MP3 requires FFmpeg",
+							JOptionPane.YES_NO_CANCEL_OPTION, JOptionPane.WARNING_MESSAGE);
+					if (result == JOptionPane.YES_OPTION)
+						overrideAndUseExe = true;
+					else if (result == JOptionPane.NO_OPTION)
+						continue; // Try again
+					else
+						return;
 				}
-				else
-				{
-					message = "File does not exist:\n" + ffExe.getAbsolutePath();
-					icon = JOptionPane.ERROR_MESSAGE;
-				}
-				int result = JOptionPane.showConfirmDialog(this, message, "Export to MP3 requires FFmpeg",
-						JOptionPane.OK_CANCEL_OPTION, icon);
-				if (result != JOptionPane.OK_OPTION)
-					return;
-			}
-			else if (!isFF(ffExe))
-			{
-				Object[] message = new Object[] {
-					"The MP3 converter you selected \"" + ffExe.getName() + "\" doesn't appear to be FFmpeg.\n"
-							+ "You can download FFmpeg from: ",
-					hyperlink,
-					"\nWould you like to use \"" + ffExe.getName() + "\" anyways?\n"
-							+ "If you choose No, you'll be prompted to locate ffmpeg.exe" };
-				int result = JOptionPane.showConfirmDialog(this, message, "Export to MP3 requires FFmpeg",
-						JOptionPane.YES_NO_CANCEL_OPTION, JOptionPane.WARNING_MESSAGE);
-				if (result == JOptionPane.YES_OPTION)
-					overrideAndUseExe = true;
-				else if (result == JOptionPane.NO_OPTION)
-					continue; // Try again
-				else
-					return;
-			}
 
-			mp3Prefs.put("ffExe", ffExe.getAbsolutePath());
+				mp3Prefs.put("ffExe", ffExe.getAbsolutePath());
+			}
 		}
-
 		ExportMp3Dialog mp3Dialog = new ExportMp3Dialog(this, ffExe, mp3Prefs, openedFile, abcInfo.getTitle(),
 				abcInfo.getComposer());
 		mp3Dialog.setIconImages(AbcPlayer.this.getIconImages());
-		mp3Dialog.addActionListener(new ActionListener()
-		{
-			@Override public void actionPerformed(ActionEvent e)
-			{
-				ExportMp3Dialog dialog = (ExportMp3Dialog) e.getSource();
-				JDialog waitFrame = new WaitDialog(AbcPlayer.this, dialog.getSaveFile());
-				waitFrame.setVisible(true);
-				new Thread(new ExportMp3TaskNew(sequencer.getSequence(), dialog, waitFrame)).start();
-			}
+		mp3Dialog.addActionListener(e -> {
+			ExportMp3Dialog dialog = (ExportMp3Dialog) e.getSource();
+			JDialog waitFrame = new WaitDialog(AbcPlayer.this, dialog.getSaveFile());
+			waitFrame.setVisible(true);
+			new Thread(new ExportMp3TaskNew(sequencer.getSequence(), dialog, waitFrame)).start();
 		});
 		mp3Dialog.setVisible(true);
 	}
 
-	private class ExportMp3TaskNew implements Runnable
-	{
+	private class ExportMp3TaskNew implements Runnable {
 		private Sequence sequence;
 		private ExportMp3Dialog mp3Dialog;
 		private JDialog waitFrame;
 
-		public ExportMp3TaskNew(Sequence sequence, ExportMp3Dialog mp3Dialog, JDialog waitFrame)
-		{
+		public ExportMp3TaskNew(Sequence sequence, ExportMp3Dialog mp3Dialog, JDialog waitFrame) {
 			this.sequence = sequence;
 			this.mp3Dialog = mp3Dialog;
 			this.waitFrame = waitFrame;
 		}
 
-		@Override public void run()
-		{
+		@Override
+		public void run() {
 			isExporting = true;
 			Exception error = null;
 			String ffExeSav = null;
 			Preferences mp3Prefs = mp3Dialog.getPreferencesNode();
-			try
-			{
-				ffExeSav = mp3Prefs.get("ffExe", null);
-				mp3Prefs.put("ffExe", "");
+			try {
+				if (System.getProperty("os.name").contains("Windows")) {
+					ffExeSav = mp3Prefs.get("ffExe", null);
+					mp3Prefs.put("ffExe", "");
+				}
 				File wavFile = File.createTempFile("AbcPlayer-", ".wav");
-				FileOutputStream fos = new FileOutputStream(wavFile);
-				try
-				{
+				try (FileOutputStream fos = new FileOutputStream(wavFile)) {
 					MidiToWav.render(sequence, fos);
 					fos.close();
-					String commando = mp3Dialog.getCommandLineNew(wavFile);
-					//System.out.println(commando);
-					Process p = Runtime.getRuntime().exec(commando);
-					if (p.waitFor() != 0)
+					String[] args = mp3Dialog.getCommandLineNew(wavFile).toArray(new String[0]);
+
+					ProcessBuilder ps = new ProcessBuilder(args);
+
+					ps.redirectErrorStream(true);
+
+					Process p = ps.start();
+
+					BufferedReader in = new BufferedReader(new InputStreamReader(p.getInputStream()));
+					String line;
+
+					while ((line = in.readLine()) != null) {
+						System.out.println(line);
+					}
+					if (p.waitFor() != 0) {
 						throw new Exception("FFmpeg failed");
-				}
-				finally
-				{
-					fos.close();
+					}
+				} finally {
 					wavFile.delete();
 				}
-			}
-			catch (Exception e)
-			{
+			} catch (Exception e) {
 				error = e;
-			}
-			finally
-			{
-				if (ffExeSav != null)
-				{
+			} finally {
+				if (ffExeSav != null && System.getProperty("os.name").contains("Windows"))
 					mp3Prefs.put("ffExe", ffExeSav);
-				}
-				isExporting = false;
-				SwingUtilities.invokeLater(new ExportMp3FinishedTask(error, waitFrame));
 			}
+			isExporting = false;
+			SwingUtilities.invokeLater(new ExportMp3FinishedTask(error, waitFrame));
 		}
 	}
 
-	private class ExportMp3FinishedTask implements Runnable
-	{
+	private class ExportMp3FinishedTask implements Runnable {
 		private Exception error;
 		private JDialog waitFrame;
 
-		public ExportMp3FinishedTask(Exception error, JDialog waitFrame)
-		{
+		public ExportMp3FinishedTask(Exception error, JDialog waitFrame) {
 			this.error = error;
 			this.waitFrame = waitFrame;
 		}
 
-		@Override public void run()
-		{
-			if (error != null)
-			{
+		@Override
+		public void run() {
+			if (error != null) {
 				JOptionPane.showMessageDialog(AbcPlayer.this, error.getMessage(), "Error saving MP3 file",
 						JOptionPane.ERROR_MESSAGE);
 			}
 			waitFrame.setVisible(false);
 		}
+
 	}
 
-	private void refreshSequence()
-	{
+	private void refreshSequence() {
 		long position = sequencer.getPosition();
 		Sequence song;
 
-		try
-		{
+		try {
 			AbcToMidi.Params params = new AbcToMidi.Params(abcData);
 			params.useLotroInstruments = useLotroInstruments;
 			params.instrumentOverrideMap = instrumentOverrideMap;
@@ -2255,90 +1852,95 @@ public class AbcPlayer extends JFrame implements TableLayoutConstants, MidiConst
 			params.enableLotroErrors = false;
 			params.stereo = stereoMenuItem.isSelected();
 			song = AbcToMidi.convert(params);
-		}
-		catch (ParseException e)
-		{
+		} catch (ParseException e) {
 			JOptionPane.showMessageDialog(this, e.getMessage(), "Error changing instrument", JOptionPane.ERROR_MESSAGE);
 			return;
 		}
 
-		try
-		{
+		try {
 			boolean running = sequencer.isRunning();
 			sequencer.reset(false);
 			sequencer.setSequence(song);
 			sequencer.setPosition(position);
 			sequencer.setRunning(running);
-		}
-		catch (InvalidMidiDataException e)
-		{
+		} catch (InvalidMidiDataException e) {
 			JOptionPane.showMessageDialog(this, e.getMessage(), "MIDI error", JOptionPane.ERROR_MESSAGE);
 			return;
 		}
 	}
-	
+
 	private static boolean openPort() {
-		
+
 		try {
-			serverSocket = new ServerSocket(9000+APP_VERSION.getBuild());
+			serverSocket = new ServerSocket(9000 + APP_VERSION.getBuild());
 			if (serverSocket == null) {
-				//System.out.println("Port is null");
+				// System.out.println("Port is null");
 				return false;
 			}
-			if (serverSocket.getLocalPort() != 9000+APP_VERSION.getBuild()) {
-				//System.out.println("Port is "+serverSocket.getLocalPort());
+			if (serverSocket.getLocalPort() != 9000 + APP_VERSION.getBuild()) {
+				// System.out.println("Port is "+serverSocket.getLocalPort());
 				return false;
-			}			
+			}
 		} catch (IOException e) {
-			//e.printStackTrace();
+			// e.printStackTrace();
 			return false;
 		}
-		//System.out.println("Made port");
-		(new Thread() {
-			public void run() {
-				try {
-					while (true) {
-						Socket socket = serverSocket.accept();
-						//System.out.println("Accepted");
-						BufferedReader in = new BufferedReader(new InputStreamReader(socket.getInputStream(), "UTF16"));
-				        
-			        	//while (socket.isConnected()) {
-			        		String data = in.readLine();
-						
-			        		if (data != null && data.length() >= 4 && data.substring(data.length() - 4).equalsIgnoreCase(".abc")) {// && !data.substring(0,3).equalsIgnoreCase("GET") && 
-			        			//System.out.println("Receiving file path ("+data.length()+" chars) from port "+(9000+APP_VERSION.getBuild())+":\n"+data);
-			        			String[] datas = {data};
-			        			activate(datas);
-			        		} else {
-			        			//System.out.println("Received nothing");
-			        		}
-			        	//}
-			        	socket.close();
-				    }
-			    } catch (IOException e) {
-			    	//e.printStackTrace();
-			    }
+		// System.out.println("Made port");
+		(new Thread(() -> {
+			try {
+				while (true) {
+					Socket socket = serverSocket.accept();
+					// System.out.println("Accepted");
+					BufferedReader in = new BufferedReader(
+							new InputStreamReader(socket.getInputStream(), StandardCharsets.UTF_16));
+
+					// while (socket.isConnected()) {
+					String data = in.readLine();
+
+					if (data != null && data.length() >= 4
+							&& data.substring(data.length() - 4).equalsIgnoreCase(".abc")) {// &&
+																							// !data.substring(0,3).equalsIgnoreCase("GET")
+																							// &&
+						// System.out.println("Receiving file path ("+data.length()+" chars) from port
+						// "+(9000+APP_VERSION.getBuild())+":\n"+data);
+						String[] datas = { data };
+						activate(datas);
+					} else {
+						// System.out.println("Received nothing");
+					}
+					// }
+					socket.close();
+				}
+			} catch (IOException e) {
+				// e.printStackTrace();
 			}
-		}).start();
-	    return true;
+		})).start();
+		return true;
 	}
-	
+
 	private static void sendArgsToPort(final String[] args) {
 		if (args == null || args.length == 0 || args[0].length() < 3) {
-			//System.out.println("AbcPlayer already running. No filepath detected. Closing.");
+			// System.out.println("AbcPlayer already running. No filepath detected.
+			// Closing.");
 			return;
 		}
-		try {			
-			Socket clientSocket = new Socket("localhost", 9000+APP_VERSION.getBuild());
-			OutputStreamWriter os = new OutputStreamWriter(clientSocket.getOutputStream(), "UTF16");//NTFS uses UTF16 for filenames
-			//for (String arg : args) {
-				os.write(args[0]);
-				os.close();//Must be here to flush to stream
-				//System.out.println("AbcPlayer already running. Sending file path ("+args[0].length()+" chars) to port "+(9000+APP_VERSION.getBuild())+":\n"+args[0]);
-			//}
+		try {
+			Socket clientSocket = new Socket("localhost", 9000 + APP_VERSION.getBuild());
+			OutputStreamWriter os = new OutputStreamWriter(clientSocket.getOutputStream(), StandardCharsets.UTF_16);// NTFS
+																													// uses
+																													// UTF16
+																													// for
+																													// filenames
+			// for (String arg : args) {
+			os.write(args[0]);
+			os.close();// Must be here to flush to stream
+			// System.out.println("AbcPlayer already running. Sending file path
+			// ("+args[0].length()+" chars) to port
+			// "+(9000+APP_VERSION.getBuild())+":\n"+args[0]);
+			// }
 			clientSocket.close();
 		} catch (IOException e) {
-			//e.printStackTrace();
+			// e.printStackTrace();
 		}
 	}
 }
