@@ -1,6 +1,7 @@
 package com.digero.common.abc;
 
 import java.io.UnsupportedEncodingException;
+import java.nio.charset.StandardCharsets;
 import java.text.Normalizer;
 
 public class StringCleaner {
@@ -33,18 +34,18 @@ public class StringCleaner {
 	}
 	
 	private static String replaceUmlaut(String before) {
-		before = before.replaceAll("�", "ss");
-		before = before.replaceAll("�", "ae");
-		before = before.replaceAll("�", "AE");
-		before = before.replaceAll("�", "oe");
-		before = before.replaceAll("�", "OE");
-		before = before.replaceAll("�", "aa");
-		before = before.replaceAll("�", "AA");
-		before = before.replaceAll("�", "ae");
-		before = before.replaceAll("�", "AE");
-		before = before.replaceAll("�", "oe");
-		before = before.replaceAll("�", "OE");
-		before = before.replaceAll("�", "-");
+		before = before.replaceAll("ß", "ss");
+		before = before.replaceAll("æ", "ae");
+		before = before.replaceAll("Æ", "AE");
+		before = before.replaceAll("ø", "oe");
+		before = before.replaceAll("Ø", "OE");
+		before = before.replaceAll("å", "aa");
+		before = before.replaceAll("Å", "AA");
+		before = before.replaceAll("ä", "ae");
+		before = before.replaceAll("Ä", "AE");
+		before = before.replaceAll("ö", "oe");
+		before = before.replaceAll("Ö", "OE");
+		before = before.replaceAll("—", "-");
 	    return before;
 	}
 	
@@ -53,12 +54,8 @@ public class StringCleaner {
 	    String regex = "[\\p{InCombiningDiacriticalMarks}\\p{IsLm}\\p{IsSk}]+";
 	    //String regex = "[\\P{InBasicLatin}]+";
 	    String after = new String(s1);
-	    try {
-	    	after = new String(s1.replaceAll(regex, "").getBytes("ascii"), "ascii");
-		} catch (UnsupportedEncodingException e) {
-			//e.printStackTrace();
-		}
-	    after = after.replace('?', ' ');
+		after = new String(s1.replaceAll(regex, "").getBytes(StandardCharsets.US_ASCII), StandardCharsets.US_ASCII);
+		after = after.replace('?', ' ');
 	    return after;
 	}
 	
